+++ R/check_plate.R
@@ -1,11 +1,19 @@
 #' Checks whether the input file is valid for use with the `tidy_plate()`
 #' function
 #'
+#' @description
+#' `check_plate()` performs quality checks on the input microwell shaped data and
+#' warns the user if there is any discrepancy. The user can either fix the input
+#' file or use the `build_plate()` function to build a template csv or xlsx file.
+#'
+#'
 #' @param file A character string containing the path to a csv or excel file.
 #' @param well_id A character string that will be the name for the well id
 #' column.
 #' @param sheet If file type is xlsx this is the sheet name (character) or
 #' number (integer).
+#'
+#' @seealso [build_plate()]
 #'
 #' @return An message indicating whether the input file is compatible with the
 #' `tidy_plate()` function
@@ -26,20 +34,20 @@ check_plate <- function(file,
   # Check whether function arguments are valid----
   ## One file should be provided----
   if (length(file) != 1) {
-    stop(
+    rlang::abort(
       paste0(
         "Invalid input: ",
         ifelse(length(file) > 1,
                "More than one file provided."
         )
       ),
-      call. = FALSE
+      call = NULL
     )
   }
 
   ## `well_id` should be a character vector of length 1----
   if (!is.character(well_id) || length(well_id) != 1L) {
-    stop("`well_id` should be a character vector of length 1", call. = FALSE)
+    rlang::abort("`well_id` should be a character vector of length 1", call = NULL)
   }
 
   ## Read file ext and basename----
@@ -48,7 +56,7 @@ check_plate <- function(file,
 
   ## Check if file exists----
   if (!(file.exists(file))) {
-    stop(paste0(file_full_name, " does not exist!"), call. = FALSE)
+    rlang::abort(paste0(file_full_name, " does not exist!"), call = NULL)
   }
 
   # Read data----
@@ -63,7 +71,7 @@ check_plate <- function(file,
                                   count_columns ,
                                   count_rows_actual,
                                   well_id,
-                                  file_full_name)[[1]]
+                                  file_full_name)
 
-  message(paste0(file_full_name, ": OK; Plate type: ", plate_parameters[[3]], "-well"))
+  rlang::inform(paste0(file_full_name, ": OK; Plate type: ", plate_parameters[[3]], "-well"))
 }